Explicitly declared all casts to remove warnings

+++ Source/Lib/Common/Codec/EbRateControlProcess.c
@@ -446,7 +446,7 @@ void high_level_rc_input_picture_vbr(
                     queue_entry_index_head_temp + HIGH_LEVEL_RATE_CONTROL_HISTOGRAM_QUEUE_MAX_DEPTH : 
                     queue_entry_index_head_temp;
 
-            queue_entry_index_temp = queue_entry_index_head_temp;
+            queue_entry_index_temp = (uint32_t) queue_entry_index_head_temp;
             {
                 hl_rate_control_histogram_ptr_temp = (encode_context_ptr->hl_rate_control_historgram_queue[queue_entry_index_temp]);
 
@@ -553,7 +553,7 @@ void high_level_rc_input_picture_vbr(
                         queue_entry_index_head_temp + HIGH_LEVEL_RATE_CONTROL_HISTOGRAM_QUEUE_MAX_DEPTH : 
                         queue_entry_index_head_temp;
 
-                queue_entry_index_temp = queue_entry_index_head_temp;
+                queue_entry_index_temp = (uint32_t) queue_entry_index_head_temp;
                 // This is set to false, so the last frame would go inside the loop
                 end_of_sequence_flag = EB_FALSE;
 
@@ -632,7 +632,7 @@ void high_level_rc_input_picture_vbr(
                         queue_entry_index_head_temp + HIGH_LEVEL_RATE_CONTROL_HISTOGRAM_QUEUE_MAX_DEPTH : 
                         queue_entry_index_head_temp;
 
-                queue_entry_index_temp = queue_entry_index_head_temp;
+                queue_entry_index_temp = (uint32_t) queue_entry_index_head_temp;
 
                 // This is set to false, so the last frame would go inside the loop
                 end_of_sequence_flag = EB_FALSE;
@@ -688,7 +688,7 @@ void high_level_rc_input_picture_vbr(
                         queue_entry_index_head_temp + HIGH_LEVEL_RATE_CONTROL_HISTOGRAM_QUEUE_MAX_DEPTH : 
                         queue_entry_index_head_temp;
 
-                queue_entry_index_temp = queue_entry_index_head_temp;
+                queue_entry_index_temp = (uint32_t) queue_entry_index_head_temp;
 
                 // This is set to false, so the last frame would go inside the loop
                 end_of_sequence_flag = EB_FALSE;
@@ -1756,7 +1756,7 @@ void high_level_rc_input_picture_cvbr(
                     queue_entry_index_head_temp + HIGH_LEVEL_RATE_CONTROL_HISTOGRAM_QUEUE_MAX_DEPTH :
                     queue_entry_index_head_temp;
 
-            queue_entry_index_temp = queue_entry_index_head_temp;
+            queue_entry_index_temp = (uint32_t) queue_entry_index_head_temp;
             {
                 hl_rate_control_histogram_ptr_temp = (encode_context_ptr->hl_rate_control_historgram_queue[queue_entry_index_temp]);
 
@@ -1866,7 +1866,7 @@ void high_level_rc_input_picture_cvbr(
                             queue_entry_index_head_temp + HIGH_LEVEL_RATE_CONTROL_HISTOGRAM_QUEUE_MAX_DEPTH :
                             queue_entry_index_head_temp;
 
-                    queue_entry_index_temp = queue_entry_index_head_temp;
+                    queue_entry_index_temp = (uint32_t) queue_entry_index_head_temp;
                     // This is set to false, so the last frame would go inside the loop
                     end_of_sequence_flag = EB_FALSE;
 
@@ -1949,7 +1949,7 @@ void high_level_rc_input_picture_cvbr(
                         queue_entry_index_head_temp + HIGH_LEVEL_RATE_CONTROL_HISTOGRAM_QUEUE_MAX_DEPTH :
                         queue_entry_index_head_temp;
 
-                queue_entry_index_temp = queue_entry_index_head_temp;
+                queue_entry_index_temp = (uint32_t) queue_entry_index_head_temp;
 
                 // This is set to false, so the last frame would go inside the loop
                 end_of_sequence_flag = EB_FALSE;
@@ -2003,7 +2003,7 @@ void high_level_rc_input_picture_cvbr(
                         queue_entry_index_head_temp + HIGH_LEVEL_RATE_CONTROL_HISTOGRAM_QUEUE_MAX_DEPTH :
                         queue_entry_index_head_temp;
 
-                queue_entry_index_temp = queue_entry_index_head_temp;
+                queue_entry_index_temp = (uint32_t) queue_entry_index_head_temp;
 
                 // This is set to false, so the last frame would go inside the loop
                 end_of_sequence_flag = EB_FALSE;
@@ -2313,7 +2313,7 @@ void frame_level_rc_input_picture_cvbr(
                 ref_qp_table_index);
             high_level_rate_control_ptr->pred_bits_ref_qpPerSw[ref_qp_index] = 0;
 
-            queue_entry_index_temp = queue_entry_index_head_temp;
+            queue_entry_index_temp = (uint32_t) queue_entry_index_head_temp;
             // This is set to false, so the last frame would go inside the loop
             end_of_sequence_flag = EB_FALSE;
